requirements.txt
pydantic==2.0.2
django
djangorestframework
psycopg2
requests